+++ src/azure-cli/azure/cli/command_modules/network/commands.py
@@ -442,7 +442,9 @@ def load_command_table(self, _):
         g.custom_command('create', 'create_ag_url_path_map_rule', supports_no_wait=True, validator=process_ag_url_path_map_rule_create_namespace)
         g.custom_command('delete', 'delete_ag_url_path_map_rule', supports_no_wait=True)
 
-    with self.command_group('network application-gateway waf-config') as g:
+    waf_config_deprecate_info = self.deprecate(redirect='network application-gateway waf-policy', hide=False)
+    with self.command_group('network application-gateway waf-config',
+                            deprecate_info=waf_config_deprecate_info) as g:
         g.custom_command('set', 'set_ag_waf_config_2017_03_01', min_api='2017-03-01', supports_no_wait=True)
         g.custom_command('set', 'set_ag_waf_config_2016_09_01', max_api='2016-09-01', supports_no_wait=True)
         g.custom_show_command('show', 'show_ag_waf_config')
@@ -455,7 +457,9 @@ def load_command_table(self, _):
     # endregion
 
     # region ApplicationGatewayWAFPolicy
-    with self.command_group('network application-gateway waf-policy', network_ag_waf_sdk, client_factory=cf_app_gateway_waf_policy, min_api='2018-12-01') as g:
+    with self.command_group('network application-gateway waf-policy', network_ag_waf_sdk,
+                            client_factory=cf_app_gateway_waf_policy,
+                            min_api='2018-12-01') as g:
         g.custom_command('create', 'create_ag_waf_policy')
         g.command('delete', 'delete')
         g.show_command('show', 'get')
@@ -463,21 +467,53 @@ def load_command_table(self, _):
         g.generic_update_command('update', custom_func_name='update_ag_waf_policy')
         g.wait_command('wait')
 
-    with self.command_group('network application-gateway waf-policy rule', network_ag_waf_sdk, client_factory=cf_app_gateway_waf_policy, min_api='2018-12-01') as g:
-        g.custom_command('create', 'create_ag_waf_rule')
-        g.custom_command('delete', 'delete_ag_waf_rule')
-        g.custom_command('list', 'list_ag_waf_rules')
-        g.custom_show_command('show', 'show_ag_waf_rule')
-        g.generic_update_command('update', command_type=network_ag_waf_sdk,
+    with self.command_group('network application-gateway waf-policy policy-setting', network_ag_waf_sdk,
+                            client_factory=cf_app_gateway_waf_policy,
+                            min_api='2019-09-01') as g:
+        g.custom_command('list', 'list_waf_policy_setting')
+        g.generic_update_command('update',
+                                 command_type=network_ag_waf_sdk,
                                  client_factory=cf_app_gateway_waf_policy,
-                                 custom_func_name='update_ag_waf_rule',
+                                 custom_func_name='update_waf_policy_setting')
+
+    with self.command_group('network application-gateway waf-policy custom-rule', network_ag_waf_sdk,
+                            client_factory=cf_app_gateway_waf_policy,
+                            min_api='2018-12-01') as g:
+        g.custom_command('create', 'create_waf_custom_rule')
+        g.custom_command('delete', 'delete_waf_custom_rule')
+        g.custom_command('list', 'list_waf_custom_rules')
+        g.custom_show_command('show', 'show_waf_custom_rule')
+        g.generic_update_command('update',
+                                 command_type=network_ag_waf_sdk,
+                                 client_factory=cf_app_gateway_waf_policy,
+                                 custom_func_name='update_waf_custom_rule',
                                  child_collection_prop_name='custom_rules',
                                  child_arg_name='rule_name')
 
-    with self.command_group('network application-gateway waf-policy rule match-condition', network_ag_waf_sdk, client_factory=cf_app_gateway_waf_policy, min_api='2018-12-01') as g:
-        g.custom_command('add', 'add_ag_waf_rule_match_cond')
-        g.custom_command('list', 'list_ag_waf_rule_match_cond')
-        g.custom_command('remove', 'remove_ag_waf_rule_match_cond')
+    with self.command_group('network application-gateway waf-policy custom-rule match-condition', network_ag_waf_sdk,
+                            client_factory=cf_app_gateway_waf_policy,
+                            min_api='2018-12-01') as g:
+        g.custom_command('add', 'add_waf_custom_rule_match_cond')
+        g.custom_command('list', 'list_waf_custom_rule_match_cond')
+        g.custom_command('remove', 'remove_waf_custom_rule_match_cond')
+
+    with self.command_group('network application-gateway waf-policy managed-rule rule-set', network_ag_waf_sdk,
+                            client_factory=cf_app_gateway_waf_policy,
+                            min_api='2019-09-01') as g:
+        g.custom_command('add', 'add_waf_managed_rule_set')
+        g.generic_update_command('update',
+                                 command_type=network_ag_waf_sdk,
+                                 client_factory=cf_app_gateway_waf_policy,
+                                 custom_func_name='update_waf_managed_rule_set')
+        g.custom_command('remove', 'remove_waf_managed_rule_set')
+        g.custom_command('list', 'list_waf_managed_rule_set')
+
+    with self.command_group('network application-gateway waf-policy managed-rule exclusion', network_ag_waf_sdk,
+                            client_factory=cf_app_gateway_waf_policy,
+                            min_api='2019-09-01') as g:
+        g.custom_command('add', 'add_waf_managed_rule_exclusion')
+        g.custom_command('remove', 'remove_waf_managed_rule_exclusion')
+        g.custom_command('list', 'list_waf_managed_rule_exclusion')
     # endregion
 
     # region ApplicationSecurityGroups
@@ -981,6 +1017,11 @@ def load_command_table(self, _):
         g.custom_command('add', 'add_vnet_gateway_ipsec_policy', supports_no_wait=True, doc_string_source='IpsecPolicy')
         g.custom_command('list', 'list_vnet_gateway_ipsec_policies')
         g.custom_command('clear', 'clear_vnet_gateway_ipsec_policies', supports_no_wait=True)
+
+    with self.command_group('network vnet-gateway aad', network_vgw_sdk, min_api='2019-04-01') as g:
+        g.custom_command('assign', 'assign_vnet_gateway_aad', supports_no_wait=True)
+        g.custom_command('show', 'show_vnet_gateway_aad')
+        g.custom_command('remove', 'remove_vnet_gateway_aad', supports_no_wait=True)
     # endregion
 
     # region VirtualNetworkGatewayConnections