+++ src/azure-cli/azure/cli/command_modules/vm/tests/latest/test_vm_commands.py
@@ -678,10 +678,11 @@ class VMManagedDiskScenarioTest(ScenarioTest):
                  ])
 
         self.cmd('image create -g {rg} -n {image_3} --source {snapshot1} --data-disk-sources {disk1} {snapshot2_id} {disk2_id}'
-                 ' --os-type Linux --tags tag1=i1 --storage-sku Standard_LRS --os-disk-caching ReadWrite',
+                 ' --os-type Linux --tags tag1=i1 --storage-sku Standard_LRS --os-disk-caching ReadWrite --data-disk-caching ReadOnly',
                  checks=[
                      self.check('storageProfile.osDisk.storageAccountType', 'Standard_LRS'),
-                     self.check('storageProfile.osDisk.caching', 'ReadWrite')
+                     self.check('storageProfile.osDisk.caching', 'ReadWrite'),
+                     self.check('storageProfile.dataDisks[0].caching', 'ReadOnly')
                  ])
 
     @ResourceGroupPreparer(name_prefix='cli_test_vm_disk_upload_')
@@ -1198,16 +1199,17 @@ class VMCreateNoneOptionsTest(ScenarioTest):  # pylint: disable=too-many-instanc
 
 class VMCreateMonitorTest(ScenarioTest):
 
-    @ResourceGroupPreparer(name_prefix='cli_test_vm_create_with_monitor', location='centralus')
+    @ResourceGroupPreparer(name_prefix='cli_test_vm_create_with_monitor', location='eastus')
     def test_vm_create_with_monitor(self, resource_group):
 
         self.kwargs.update({
             'vm': 'monitorvm',
-            'workspace': 'vmlogworkspace20191009',
+            'workspace': self.create_random_name('cliworkspace', 20),
             'rg': resource_group
         })
 
         self.cmd('vm create -n {vm} -g {rg} --image UbuntuLTS --workspace {workspace}')
+        self.cmd('vm monitor log show -n {vm} -g {rg} -q "Perf | limit 10"')
 
 
 class VMBootDiagnostics(ScenarioTest):
@@ -1425,7 +1427,7 @@ class VMCreateExistingOptions(ScenarioTest):
             'vmss': 'vmss1'
         })
 
-        self.cmd('vmss create -g {rg} -n {vmss} --orchestration-mode VM')
+        self.cmd('vmss create -g {rg} -n {vmss} --orchestration-mode VM --platform-fault-domain-count 2')
         self.cmd('vm create -g {rg} -n {vm} --image ubuntults --vmss {vmss}')
         vmss_id = self.cmd('vmss show -g {rg} -n {vmss}').get_output_in_json()['id']
         self.cmd('vm show -g {rg} -n {vm}', checks=[
@@ -3304,6 +3306,18 @@ class VMGalleryImage(ScenarioTest):
         self.cmd('sig image-definition delete -g {rg} --gallery-name {gallery} --gallery-image-definition {image}')
         self.cmd('sig delete -g {rg} --gallery-name {gallery}')
 
+    @ResourceGroupPreparer(name_prefix='cli_test_gallery_specialized_', location='eastus2')
+    def test_gallery_specialized(self, resource_group):
+        self.kwargs.update({
+            'gallery': 'gallery1',
+            'image': 'image1'
+        })
+        self.cmd('sig create -g {rg} --gallery-name {gallery}', checks=self.check('name', '{gallery}'))
+        self.cmd('sig image-definition create -g {rg} --gallery-name {gallery} --gallery-image-definition {image} --os-type linux --os-state specialized --hyper-v-generation V2 -p publisher1 -f offer1 -s sku1',
+                 checks=[self.check('name', '{image}'), self.check('osState', 'Specialized'),
+                         self.check('hyperVgeneration', 'V2')])
+
+
 # endregion
 
 
@@ -3540,7 +3554,7 @@ class VMPriorityEvictionBillingTest(ScenarioTest):
         })
 
         # vm create
-        self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --priority Low --eviction-policy Deallocate --max-billing=50')
+        self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --priority Low --eviction-policy Deallocate --max-price 50')
 
         self.cmd('vm show -g {rg} -n {vm}', checks=[
             self.check('priority', 'Low'),
@@ -3549,7 +3563,7 @@ class VMPriorityEvictionBillingTest(ScenarioTest):
         ])
 
         # vmss create
-        self.cmd('vmss create -g {rg} -n {vmss} --image UbuntuLTS --lb-sku Standard --priority Low --eviction-policy Deallocate --max-billing=50', checks=[
+        self.cmd('vmss create -g {rg} -n {vmss} --image UbuntuLTS --lb-sku Standard --priority Low --eviction-policy Deallocate --max-price 50', checks=[
             self.check('vmss.virtualMachineProfile.priority', 'Low'),
             self.check('vmss.virtualMachineProfile.evictionPolicy', 'Deallocate'),
             self.check('vmss.virtualMachineProfile.billingProfile.maxPrice', 50)
@@ -3612,6 +3626,23 @@ class VMImageTermsTest(ScenarioTest):
         self.cmd('vm image terms show --publisher {publisher} --offer {offer} --plan {plan}', checks=[
             self.check('accepted', False)
         ])
+
+
+class VMSSOrchestrationModeTest(ScenarioTest):
+
+    @ResourceGroupPreparer(name_prefix='cli_test_vmss_orchestration_mode_', location='eastus')
+    def test_vmss_orchestration_mode(self, resource_group):
+        self.kwargs.update({
+            'vmss': 'vmss1',
+            'vmss2': 'vmss2'
+        })
+
+        self.cmd('vmss create -g {rg} -n {vmss} --orchestration-mode VM --zones 3 --platform-fault-domain-count 5')
+        self.cmd('vmss show -g {rg} -n {vmss}', checks=[
+            self.check('name', '{vmss}')
+        ])
+        with self.assertRaises(CLIError):
+            self.cmd('vmss create -g {rg} -n {vmss2} --orchestration-mode VM --admin-username user --admin-password 123456')
 
 
 if __name__ == '__main__':